backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain_google_genai
langchain_core
langchain_groq
langchain_community
pydantic
python-dotenv
python-multipart
python-pptx
pydub
PyMuPDF
typing-extensions
uuid
frontend
appwrite
murf